api protection

+++ src/app/api/admin/promoCode/route.ts
@@ -1,11 +1,23 @@
 import {connect} from "@/db/db";
 import PromoCode from "@/models/promoCodeModel";
 import {NextRequest, NextResponse} from "next/server";
+import {currentUser, isAdmin} from "@/lib/auth";
 connect()
 
 
 export async function GET(request: NextRequest) {
     try {
+
+        const user = await currentUser();
+
+        if(!user){
+            return NextResponse.json({error: "Unauthorized."}, {status: 401})
+        }
+
+        if(user?.isAdmin === false){
+            return NextResponse.json({error: "Forbidden. You don't have administrator rights."}, {status: 403})
+        }
+
         const allPromoCodes = await PromoCode.find({});
         return NextResponse.json({
             message: "All PromoCodes retrieved successfully",
@@ -18,6 +30,17 @@ export async function GET(request: NextRequest) {
 
 export async function POST(request: NextRequest) {
     try {
+
+        const user = await currentUser();
+
+        if(!user){
+            return NextResponse.json({error: "Unauthorized."}, {status: 401})
+        }
+
+        if(user?.isAdmin === false){
+            return NextResponse.json({error: "Forbidden. You don't have administrator rights."}, {status: 403})
+        }
+
         const reqBody = await request.json()
         const {title, value, isValid} = reqBody
         const newPromoCode = new PromoCode({title, value, isValid});
@@ -33,6 +56,17 @@ export async function POST(request: NextRequest) {
 
 export async function DELETE(request: NextRequest) {
     try {
+
+        const user = await currentUser();
+
+        if(!user){
+            return NextResponse.json({error: "Unauthorized."}, {status: 401})
+        }
+
+        if(user?.isAdmin === false){
+            return NextResponse.json({error: "Forbidden. You don't have administrator rights."}, {status: 403})
+        }
+
         const reqBody = await request.json()
         const { promoCodeId } = reqBody
         const deletedPromoCode = await PromoCode.findByIdAndDelete(promoCodeId)
@@ -50,6 +84,17 @@ export async function DELETE(request: NextRequest) {
 
 export async function PUT(request: NextRequest) {
     try {
+
+        const user = await currentUser();
+
+        if(!user){
+            return NextResponse.json({error: "Unauthorized."}, {status: 401})
+        }
+
+        if(user?.isAdmin === false){
+            return NextResponse.json({error: "Forbidden. You don't have administrator rights."}, {status: 403})
+        }
+
         const reqBody = await request.json()
         const { promoCodeId, isValid } = reqBody
         const promoCode = await PromoCode.findById(promoCodeId)
