+++ src/app/api/users/messages/route.ts
@@ -1,11 +1,23 @@
 import {connect} from "@/db/db";
 import Messages from "@/models/messageModel";
 import {NextRequest, NextResponse} from "next/server";
+import {currentUser} from "@/lib/auth";
 connect()
 
 
 export async function GET(request: NextRequest) {
     try {
+
+        const user = await currentUser();
+
+        if(!user){
+            return NextResponse.json({error: "Unauthorized."}, {status: 401})
+        }
+
+        if(user?.isAdmin === false){
+            return NextResponse.json({error: "Forbidden. You don't have administrator rights."}, {status: 403})
+        }
+
         const allMessages = await Messages.find({});
         return NextResponse.json({
             message: "All Messages retrieved successfully",
@@ -18,6 +30,7 @@ export async function GET(request: NextRequest) {
 
 export async function POST(request: NextRequest) {
     try {
+
         const reqBody = await request.json()
         const {title, message, category, authorsContact, createdAt} = reqBody
         const newMessage = new Messages({title, message, category, authorsContact, createdAt});
@@ -33,6 +46,17 @@ export async function POST(request: NextRequest) {
 
 export async function DELETE(request: NextRequest) {
     try {
+
+        const user = await currentUser();
+
+        if(!user){
+            return NextResponse.json({error: "Unauthorized."}, {status: 401})
+        }
+
+        if(user?.isAdmin === false){
+            return NextResponse.json({error: "Forbidden. You don't have administrator rights."}, {status: 403})
+        }
+
         const reqBody = await request.json()
         const { messageId } = reqBody
         const deletedMessage = await Messages.findByIdAndDelete(messageId)
